Add rails 3 compatibility
