+++ .travis.yml
@@ -2,8 +2,9 @@ language: ruby
 cache: bundler
 matrix:
   include:
+    - rvm: 2.2.1
+      gemfile: gemfiles/rails32.gemfile
     - rvm: 2.3.0
       gemfile: gemfiles/rails42.gemfile
     - rvm: 2.3.1
       gemfile: gemfiles/rails5.gemfile
-
