+++ gemfiles/rails32.gemfile
@@ -1,0 +1,7 @@
+source 'https://rubygems.org'
+
+gem 'sqlite3', platform: :mri
+gem 'activerecord-jdbcsqlite3-adapter', platform: :jruby
+gem 'activerecord', '~> 3.2.22'
+
+gemspec path: '..'
