+++ influxer.gemspec
@@ -15,7 +15,7 @@ Gem::Specification.new do |s|
   s.files         = `git ls-files`.split($/)
   s.require_paths = ["lib"]
 
-  s.add_dependency "activemodel", '>= 4.0'
+  s.add_dependency "activemodel", '>= 3.2.0'
   s.add_dependency "influxdb", "~> 0.3"
   s.add_dependency "anyway_config", "~>0.3.0"
 
@@ -23,7 +23,7 @@ Gem::Specification.new do |s|
   s.add_development_dependency "simplecov", ">= 0.3.8"
   s.add_development_dependency 'rake', '~> 10.1'
   s.add_development_dependency 'sqlite3'
-  s.add_development_dependency 'activerecord', '>= 4.0'
+  s.add_development_dependency 'activerecord', '>= 3.2.0'
   s.add_development_dependency 'pry-byebug'
   s.add_development_dependency "rspec", ">= 3.1.0"
   s.add_development_dependency "webmock", "~> 1.21.0"
